requirements.txt
customtkinter
pillow
numpy
tensorflow
tf_keras
